+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added table to field description cell to show enumerated values ([#123](https://github.jpl.nasa.gov/regvue/regvue/pull/123)).
 - Added support for 16 bit registers ([#110](https://github.jpl.nasa.gov/regvue/regvue/pull/110)).
 ### Fixed
 - Only display a single bit for the bit range of single bit fields ([#120](https://github.jpl.nasa.gov/regvue/regvue/pull/120)).